% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{alwdeff}
\alias{alwdeff}
\title{Always defective strategy}
\usage{
alwdeff(...)
}
\arguments{
\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 2 for NO cooperation
}
\description{
Always defective strategy
}
\examples{
nextmove<-alwdeff()

}
