% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{alwfunc}
\alias{alwfunc}
\title{Always functional strategy}
\usage{
alwfunc(...)
}
\arguments{
\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 1 for cooperation
}
\description{
Always functional strategy
}
