% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{check_strategy}
\alias{check_strategy}
\title{Check strategy}
\usage{
check_strategy(strategy = "Alwaysfunctional")
}
\arguments{
\item{strategy}{String: must be one of the following:
'Alwaysfunctional', 'Alwaysdefectiveinterfering',
'Randommodification', 'Randomeffective',
'Randomdefectiveinterfering' or 'Count_defective'}
}
\value{
function: strategy to define next movement
}
\description{
Check strategy
}
