% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{count_def_invert}
\alias{count_def_invert}
\title{Particle counter strategy inverted}
\usage{
count_def_invert(prev, Score, k, delta, ...)
}
\arguments{
\item{prev}{logical: previous population move}

\item{Score}{Numeric: previous condition}

\item{k}{Numeric: system capacity}

\item{delta}{Numeric: factor}

\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 1 or 2 for cooperation or not.
}
\description{
Particle counter strategy inverted
}
\examples{
nextmove<-count_def_invert(prev=1,
Score=0.5,
k=1,
delta=0.3,
)
nextmove
}
