% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisoner_functions.R
\name{game}
\alias{game}
\title{game simulation}
\usage{
game(
  type = c("May", "Lotka", "Custom"),
  funGrowth,
  play1 = alwdeff,
  play2 = alwdeff,
  interaction,
  input1 = NULL,
  input2 = NULL,
  generations = NULL,
  k1 = NULL,
  k2 = NULL,
  parameters
)
}
\arguments{
\item{type}{string: method for which to run the function: \code{"May"}, \code{"Lotka"}, or
\code{"Custom"} (Must be set).}

\item{funGrowth}{function: simulates the growth dynamics.}

\item{play1}{function: simulates the population 1 decision making (defaults to always cooperative).}

\item{play2}{function: simulates the population 2 decision making (defaults to always cooperative).}

\item{interaction}{function: handles the changes in growth parameters passed to
\code{funGrowth} according to the interaction matrix.}

\item{input1}{numeric: the initial condition of population 1 (defaults to 0.1).}

\item{input2}{numeric: the initial condition of population 2 (defaults to 0.1).}

\item{generations}{numeric: the number of rounds (generations) to simulate.}

\item{k1}{numeric: the maximum system capacity for population 1. Only relevant for Lotka method (default=1).}

\item{k2}{numeric: the maximum system capacity for population 2. Only relevant for Lotka method (default=1).}

\item{parameters}{list: all the arguments ALL the necessary for \code{funGrowth}, \code{interaction}, \code{play1} and \code{play2} functions.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with the obtained scores. The output has the following columns:
}
\description{
\code{game()} returns a data.frame with the Fst values for the simulated
interactions between two populations of viral particles.
}
\examples{

#' params <- list(mutation1 = 10**-6,
#'                mutation2 = 10**-6,
#'                genome1 = 5000,
#'                genome2 = 8500,
#'                #(1:1) cooperate/cooperate
#'                r1_1_1=0.9,
#'                r2_1_1=0.19,
#'                a12_1_1=0.012,
#'                a21_1_1=-0.2,
#'                #(1:2) cooperate/defective
#'                r1_1_2=0.1,
#'                r2_1_2=0.0018,
#'                a12_1_2=-1,
#'                a21_1_2=0.1,
#'                #(2:1) defective/cooperate
#'                r1_2_1=0.15,
#'                r2_2_1=0.18,
#'                a12_2_1=-1,
#'                a21_2_1=-0.001,
#'                #(2:2) defective/defective
#'                r1_2_2=0.8,
#'                r2_2_2=-0.8,
#'                a12_2_2=-0.1,
#'                a21_2_2=-0.1)

#' simulationLV <- game(type="Lotka",
#'                      play1=count_def, #strategy of player 1#
#'                      play2=count_def, #strategy of player 2,#
#'                      parameters=params #list of parameters to pass to ode solver
#' )
}
