% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parameters.R
\name{HCV_parameters}
\alias{HCV_parameters}
\title{HCV Alwaysfunctional Alwaysdefectiveinterfering}
\usage{
HCV_parameters()
}
\value{
list: all neccesary parameters for the examples simulations
}
\description{
HCV Alwaysfunctional Alwaysdefectiveinterfering
}
\examples{
parameters <- HCV_parameters()
}
