% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction_dynamic_MAY}
\alias{interaction_dynamic_MAY}
\title{May dynamic interaction}
\usage{
interaction_dynamic_MAY(p1Move, p2Move, parameters)
}
\arguments{
\item{p1Move}{numeric: 1 or 2 for previous interaction}

\item{p2Move}{numeric: 1 or 2 for previous interaction}

\item{parameters}{list of parameters passed from the \code{game()} function}
}
\value{
A list with the modified parameters.
}
\description{
Preloaded function that handles the possible interactions
for a simulation
}
