% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNested_list}
\alias{isNested_list}
\title{Check if list is nested}
\usage{
isNested_list(l)
}
\arguments{
\item{l}{list of lists}
}
\value{
logical: True for nested list.
}
\description{
Internal function to test if a list is nested. Used for plotting
several set of parameters at once.
}
