% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{LV}
\alias{LV}
\title{Lotka-Volterra Growth function}
\usage{
LV(p1Score, p2Score, parameters)
}
\arguments{
\item{p1Score}{numeric: previous condition for population 1}

\item{p2Score}{numeric: previous condition for population 2}

\item{parameters}{list: parameters passed from the \code{game()} function}
}
\value{
numeric: the next generation values
}
\description{
Lotka-Volterra Growth function
}
\examples{
#' params <- prepare_parameters()
#'
#'
#' LV(p1Score=0.2, p2Score=0.5, parameters=params)
#'
#'
}
