% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_functions.R
\name{MAY}
\alias{MAY}
\title{May Logistic Growth function}
\usage{
MAY(p1Score, p2Score, parameters)
}
\arguments{
\item{p1Score}{numeric: previous condition for population 1}

\item{p2Score}{numeric: previous condition for population 2}

\item{parameters}{list: parameters passed from the \code{game()} function}
}
\value{
Two numerics with the next generation value
}
\description{
May Logistic Growth function
}
\examples{
#' params <- prepare_parameters()
#'
#'
#' MAY(p1Score=0.2, p2Score=0.5, parameters=params)
#'
#'
}
