% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisoner_functions.R
\name{plot_absolute}
\alias{plot_absolute}
\title{Plot absolute Fst function}
\usage{
plot_absolute(df)
}
\arguments{
\item{df}{a dataframe created with the \code{game()} function}
}
\value{
A ggplot2 plot
}
\description{
\code{plot_absolute()} is a simple helper function that returns a ggplot object
visualizing Absolute Fst across the generations. To use after running  the
\code{game()} function
}
\examples{
inst/examples/plot_absolute.R



}
