% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_relative}
\alias{plot_relative}
\title{Plot Relative  Fst function}
\usage{
plot_relative(df)
}
\arguments{
\item{df}{dataframe: created with the \code{game()} function}
}
\value{
ggplot2 object
}
\description{
\code{plot_relative()} is a simple helper function that returns a ggplot object
visualizing Absolute Fst across the generations. To use after running  the
\code{game()} function
}
\examples{

#' params <- prepare_parameters()
#'
#' simulationLV <- game(type="Lotka",
#'                      play1="Count_defective", #strategy of player 1#
#'                      play2="Count_defective", #strategy of player 2,#
#'                      parameters=params #list of parameters to pass to ode solver
#' )
#'
#'plot_relative(simulationLV)
}
