% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parameters.R
\name{prepare_parameters}
\alias{prepare_parameters}
\title{Prepare parameters}
\usage{
prepare_parameters()
}
\value{
list: all neccesary parameters for the examples simulations
}
\description{
Simple helper function to create a list of example parameter
to set up the examples simulations.
}
\examples{
parameters <- prepare_parameters()
}
