% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{randbasic}
\alias{randbasic}
\title{Random basic strategy}
\usage{
randbasic(...)
}
\arguments{
\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 1 or 2 for cooperation or not.
}
\description{
Random basic strategy
}
\examples{
nextmove<-randbasic()

}
