% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{randdeff}
\alias{randdeff}
\title{Random deffective strategy}
\usage{
randdeff(...)
}
\arguments{
\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 1 or 2 for cooperation or not.
}
\description{
Inverts Randeff
}
