% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{randeff}
\alias{randeff}
\title{Random Effective function}
\usage{
randeff(...)
}
\arguments{
\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 1 or 2 for cooperation or not.
}
\description{
Random Effective function
}
\examples{
nextmove<-randeff()

}
