% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{count_def}
\alias{count_def}
\title{Particle counter strategy}
\usage{
count_def(prev, score, k, delta, ...)
}
\arguments{
\item{prev}{Numeric: previous decision.}

\item{score}{Numeric: previous condition.}

\item{k}{Numeric: the maximum system capacity for population}

\item{delta}{Numeric: proportion of k to evaluate score. (score >= k * delta)}

\item{...}{list: extra parameters passed from \code{game()} function}
}
\value{
Numeric: 1 or 2 for cooperation or not.
}
\description{
Particle counter strategy
}
\examples{
nextmove<-count_def(prev=1,score=0.5,k=1,delta=0.5)
nextmove
}
