% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisoner_functions.R
\name{game}
\alias{game}
\title{game simulation}
\usage{
game(
  type = c("May", "Lotka", "Custom"),
  funGrowth,
  play1 = c("Alwaysfunctional", "Alwaysdefectiveinterfering", "Randommodification",
    "Randomeffective", "Randomdefectiveinterfering", "Count_defective"),
  play2 = c("Alwaysfunctional", "Alwaysdefectiveinterfering", "Randommodification",
    "Randomeffective", "Randomdefectiveinterfering", "Count_defective"),
  interaction,
  input1 = NULL,
  input2 = NULL,
  generations = NULL,
  k1 = NULL,
  k2 = NULL,
  parameters
)
}
\arguments{
\item{type}{string: method for which to run the function: \code{"May"}, \code{"Lotka"}, or
\code{"Custom"} (Must be set).}

\item{funGrowth}{function: simulates the growth dynamics.}

\item{play1}{function: simulates the population 1 decision making (defaults to always cooperative).}

\item{play2}{function: simulates the population 2 decision making (defaults to always cooperative).}

\item{interaction}{function: handles the changes in growth parameters passed to
\code{funGrowth} according to the interaction matrix.}

\item{input1}{numeric: the initial condition of population 1 (defaults to 0.1).}

\item{input2}{numeric: the initial condition of population 2 (defaults to 0.1).}

\item{generations}{numeric: the number of rounds (generations) to simulate.}

\item{k1}{numeric: the maximum system capacity for population 1. Only relevant for Lotka method (default=1).}

\item{k2}{numeric: the maximum system capacity for population 2. Only relevant for Lotka method (default=1).}

\item{parameters}{list: all the arguments ALL the necessary for \code{funGrowth}, \code{interaction}, \code{play1} and \code{play2} functions.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with the obtained scores. The output has the following columns:
}
\description{
\code{game()} returns a data.frame with the Fst values for the simulated
interactions between two populations of viral particles.
}
\examples{
#' params <- prepare_parameters()
#'
#' simulationLV <- game(type="Lotka",
#'                      play1=count_def, #strategy of player 1#
#'                      play2=count_def, #strategy of player 2,#
#'                      parameters=params #list of parameters to pass to ode solver
#' )
}
