% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisoner_functions.R
\name{plot_absolute}
\alias{plot_absolute}
\title{Plot absolute Fst function}
\usage{
plot_absolute(df)
}
\arguments{
\item{df}{a dataframe created with the \code{game()} function}
}
\value{
A ggplot2 plot
}
\description{
\code{plot_absolute()} is a simple helper function that returns a ggplot object
visualizing Absolute Fst across the generations. To use after running  the
\code{game()} function
}
\examples{

#' params <- list(mutation1 = 10**-6,
#'                mutation2 = 10**-6,
#'                genome1 = 5000,
#'                genome2 = 8500,
#'                #(1:1) cooperate/cooperate
#'                r1_1_1=0.9,
#'                r2_1_1=0.19,
#'                a12_1_1=0.012,
#'                a21_1_1=-0.2,
#'                #(1:2) cooperate/defective
#'                r1_1_2=0.1,
#'                r2_1_2=0.0018,
#'                a12_1_2=-1,
#'                a21_1_2=0.1,
#'                #(2:1) defective/cooperate
#'                r1_2_1=0.15,
#'                r2_2_1=0.18,
#'                a12_2_1=-1,
#'                a21_2_1=-0.001,
#'                #(2:2) defective/defective
#'                r1_2_2=0.8,
#'                r2_2_2=-0.8,
#'                a12_2_2=-0.1,
#'                a21_2_2=-0.1)

#' simulationLV <- game(type="Lotka",
#'                      play1=count_def, #strategy of player 1#
#'                      play2=count_def, #strategy of player 2,#
#'                      parameters=params #list of parameters to pass to ode solver
#' )


#'plot_absolute(simulationLV)
}
